+++ src/Namicorn.test.ts
@@ -225,13 +225,13 @@ describe('ZNS', () => {
     const secondEye = jest
       .spyOn(namicorn.zns, '_getResolverRecords')
       .mockResolvedValue({
-        "crypto.BCH.address": 'qrq4sk49ayvepqz7j7ep8x4km2qp8lauvcnzhveyu6',
-        "crypto.BTC.address": '1EVt92qQnaLDcmVFtHivRJaunG2mf2C3mB',
-        "crypto.ETH.address": '0x45b31e01AA6f42F0549aD482BE81635ED3149abb',
-        "crypto.LTC.address": 'LetmswTW3b7dgJ46mXuiXMUY17XbK29UmL',
-        "crypto.ZIL.address": 'zil1yu5u4hegy9v3xgluweg4en54zm8f8auwxu0xxj',
-        "ipfs.html.value": 'QmVaAtQbi3EtsfpKoLzALm6vXphdi2KjMgxEDKeGg6wHuK',
-        "ipfs.redirect_domain.value": 'www.unstoppabledomains.com',
+        'crypto.BCH.address': 'qrq4sk49ayvepqz7j7ep8x4km2qp8lauvcnzhveyu6',
+        'crypto.BTC.address': '1EVt92qQnaLDcmVFtHivRJaunG2mf2C3mB',
+        'crypto.ETH.address': '0x45b31e01AA6f42F0549aD482BE81635ED3149abb',
+        'crypto.LTC.address': 'LetmswTW3b7dgJ46mXuiXMUY17XbK29UmL',
+        'crypto.ZIL.address': 'zil1yu5u4hegy9v3xgluweg4en54zm8f8auwxu0xxj',
+        'ipfs.html.value': 'QmVaAtQbi3EtsfpKoLzALm6vXphdi2KjMgxEDKeGg6wHuK',
+        'ipfs.redirect_domain.value': 'www.unstoppabledomains.com',
       });
 
     const result = await namicorn.zns.resolution('brad.zil');
@@ -265,11 +265,11 @@ describe('ZNS', () => {
     const secondEye = jest
       .spyOn(namicorn.zns, '_getResolverRecords')
       .mockResolvedValue({
-        "ipfs.html.hash": 'QmefehFs5n8yQcGCVJnBMY3Hr6aMRHtsoniAhsM1KsHMSe',
-        "ipfs.html.value": 'QmVaAtQbi3EtsfpKoLzALm6vXphdi2KjMgxEDKeGg6wHuK',
-        "ipfs.redirect_domain.value": 'www.unstoppabledomains.com',
-        "whois.email.value": 'matt+test@unstoppabledomains.com' ,
-        "whois.for_sale.value": 'true' ,
+        'ipfs.html.hash': 'QmefehFs5n8yQcGCVJnBMY3Hr6aMRHtsoniAhsM1KsHMSe',
+        'ipfs.html.value': 'QmVaAtQbi3EtsfpKoLzALm6vXphdi2KjMgxEDKeGg6wHuK',
+        'ipfs.redirect_domain.value': 'www.unstoppabledomains.com',
+        'whois.email.value': 'matt+test@unstoppabledomains.com',
+        'whois.for_sale.value': 'true',
       });
 
     const result = await namicorn.zns.resolution('ergergergerg.zil');
@@ -297,7 +297,7 @@ describe('ZNS', () => {
     expect(zns).toBeDefined();
     const result = await zns.resolution('invalid.domain');
     expect(result).toEqual({});
-  })
+  });
 
   it('should resolve with resolution key setuped #4', async () => {
     const namicorn = new Namicorn();
@@ -305,20 +305,17 @@ describe('ZNS', () => {
     expect(zns).toBeDefined();
     const result = await zns.resolution('mcafee2020.zil');
     expect(result).toEqual({
-      crypto:
-      {
+      crypto: {
         BTC: { address: '17LV6fxL8b1pJomn5zoDR3ZCnbt88ehGBf' },
         ETH: { address: '0x0ed6180ef7c638064b9b17ff53ba76ec7077dd95' },
-        LTC: { address: 'MTbeoMfWqEZaaZVG1yE1ENoxVGNmMAxoEj' }
+        LTC: { address: 'MTbeoMfWqEZaaZVG1yE1ENoxVGNmMAxoEj' },
       },
-      whois:
-      {
+      whois: {
         email: { value: 'jordanb_970@hotmail.com' },
-        for_sale: { value: 'true' }
-      }
+        for_sale: { value: 'true' },
+      },
     });
-  })
-
+  });
 });
 
 describe('ENS', () => {