+++ src/Namicorn.ts
@@ -1,12 +1,15 @@
-import fetch from 'node-fetch';
+import nodeFetch from 'node-fetch';
 import Ens from './Ens';
 import Zns from './Zns';
 import { Blockchain, NamicornResolution } from './types';
 
 const DefaultUrl = 'https://unstoppabledomains.com/api/v1';
 
-// Node env has special properties stored in process which are not inside the browser env.
-// Multiple checks is to avoid hitting the undefined while going deeper.
+/**
+ * @ignore
+ * Node env has special properties stored in process which are not inside the browser env.
+ * Multiple checks is to avoid hitting the undefined while going deeper.
+ */
 const isNode = () => {
   if (typeof process === 'object') {
     if (typeof process.versions === 'object') {
@@ -17,6 +20,20 @@ const isNode = () => {
   }
   return false;
 };
+
+/** @ignore */
+const myfetch = isNode() ? nodeFetch : window.fetch;
+
+/** @ignore Used internaly to set the right user-agent for fetch */
+const DefaultUserAgent = isNode()
+  ? 'node-fetch/1.0 (+https://github.com/bitinn/node-fetch)'
+  : navigator.userAgent;
+/** @ignore */
+const version = require('../package.json').version;
+/** @ignore */
+const CustomUserAgent = `${DefaultUserAgent} namicorn/${version}`;
+/** @ignore */
+const headers = { 'X-user-agent': CustomUserAgent };
 
 /**
  * Blockchain domain resolution library - Namicorn.
@@ -88,9 +105,10 @@ class Namicorn {
     if (this.blockchain) {
       return await this.resolveUsingBlockchain(domain);
     } else {
-      const response = isNode()
-        ? await fetch(`${this.api}/${domain}`)
-        : await window.fetch(`${this.api}/${domain}`);
+      const response = await myfetch(`${this.api}/${domain}`, {
+        method: 'GET',
+        headers: headers,
+      });
       return response.json();
     }
   }
@@ -146,7 +164,9 @@ class Namicorn {
   /**
    * @ignore
    */
-  private async resolveUsingBlockchain(domain: string): Promise<NamicornResolution>{
+  private async resolveUsingBlockchain(
+    domain: string,
+  ): Promise<NamicornResolution> {
     const method = this.getNamingMethod(domain);
     if (!method) return null;
     const result = await method.resolve(domain);